+++ static/__tests__/iframe/iframe-utils.js
@@ -1,8 +1,7 @@
 import _ from "lodash";
 
-import { ReactColumnMenu as ColumnMenu } from "../../dtale/iframe/ColumnMenu";
-
 function findColMenuButton(result, name, btnTag = "button") {
+  const ColumnMenu = require("../../dtale/iframe/ColumnMenu").ReactColumnMenu;
   return result
     .find(ColumnMenu)
     .find(`ul li ${btnTag}`)
@@ -16,12 +15,14 @@ function clickColMenuButton(result, name, btnTag = "button") {
 }
 
 function clickColMenuSubButton(result, label, row = 0) {
+  // need to import this component here because it is using fetcher.js which gets mocked during run-time
+  const ColumnMenu = require("../../dtale/iframe/ColumnMenu").ReactColumnMenu;
   result
     .find(ColumnMenu)
     .find("ul li div.column-sorting")
     .at(row)
     .find("button")
-    .findWhere(b => _.includes(b.text(), label))
+    .findWhere(b => _.includes(b.html(), label))
     .first()
     .simulate("click");
 }