+++ static/popups/Popup.jsx
@@ -12,6 +12,7 @@ import { Describe } from "./Describe";
 import { Histogram } from "./Histogram";
 import Instances from "./Instances";
 import { Charts } from "./charts/Charts";
+import { CreateColumn } from "./create/CreateColumn";
 
 class ReactPopup extends React.Component {
   constructor(props) {
@@ -39,7 +40,7 @@ class ReactPopup extends React.Component {
           <ModalTitle>
             <i className="ico-equalizer" />
             {" Histogram for "}
-            <strong>{title}</strong>
+            <strong>{chartData.selectedCol}</strong>
             <div id="describe" />
           </ModalTitle>
         );
@@ -62,6 +63,15 @@ class ReactPopup extends React.Component {
           </ModalTitle>
         );
         body = <Describe />;
+        break;
+      case "build":
+        modalTitle = (
+          <ModalTitle>
+            <i className="ico-build" />
+            <strong>{"Build Column"}</strong>
+          </ModalTitle>
+        );
+        body = <CreateColumn />;
         break;
       case "about":
         modalTitle = (
@@ -101,6 +111,7 @@ class ReactPopup extends React.Component {
           onRequestHide: onClose,
           size: size || "modal-lg",
           backdrop: backdrop || false,
+          className: `${type}-modal`,
         }}>
         <ModalHeader>
           {modalTitle}
@@ -121,6 +132,7 @@ ReactPopup.propTypes = {
     title: PropTypes.string,
     size: PropTypes.string,
     backdrop: PropTypes.bool,
+    selectedCol: PropTypes.string,
   }),
   propagateState: PropTypes.func,
 };